+++ of_to_opml.py
@@ -1,4 +1,4 @@
-from treemodel import traverse_list, traverse_list, Visitor
+from treemodel import traverse_list, Visitor
 from omnifocus import build_model, find_database
 import os
 import codecs