+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## 1.0.0 (2013-04-09) ##
 
-- Added task flag include/exclude options --tfi, --tfe
+- Added task filters for flagged, start and due
+- Added project filters for flagged, start, completion and due
+- Reworking date logic to be more human allow "next tuesday", "last week" etc.